+++ plugin-dev/Source/Sentry/Private/GenericPlatform/Infrastructure/GenericPlatformSentryConverters.cpp
@@ -256,6 +256,20 @@ sentry_crash_reporting_mode_t FGenericPlatformSentryConverters::CrashReportingMo
 	}
 }
 
+sentry_thread_stackwalk_mode_t FGenericPlatformSentryConverters::ThreadStackwalkModeToNative(ESentryThreadStackwalkMode mode)
+{
+	switch (mode)
+	{
+	case ESentryThreadStackwalkMode::CrashedThreadOnly:
+		return SENTRY_THREAD_STACKWALK_MODE_CRASHED_ONLY;
+	case ESentryThreadStackwalkMode::AllThreads:
+		return SENTRY_THREAD_STACKWALK_MODE_ALL;
+	default:
+		UE_LOG(LogSentrySdk, Warning, TEXT("Unknown thread stackwalk mode value used. AllThreads will be returned."));
+		return SENTRY_THREAD_STACKWALK_MODE_ALL;
+	}
+}
+
 ESentryLevel FGenericPlatformSentryConverters::SentryLevelToUnreal(sentry_value_t level)
 {
 	FString levelStr = FString(UTF8_TO_TCHAR(sentry_value_as_string(level)));